+++ packages/flutter_nucleus/lib/src/build_context.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_nucleus/flutter_nucleus.dart';
 
 extension NucleusBuildContextExt on BuildContext {