+++ packages/nucleus/benchmark/benchmark/run.dart
@@ -3,6 +3,8 @@ import 'nucleus_benchmark.dart' as nucleus;
 import 'riverpod_benchmark.dart' as riverpod;
 
 void main(List<String> args) {
+  assert(false);
+
   creator.main();
   print('---');
   nucleus.main();